requirements.txt
aioredis==1.3.1
aiosmtpd==1.4.2
asgiref==3.4.1
async-timeout==4.0.2
atpublic==2.3
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.9
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.10
django-braces==1.15.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
fontawesomefree==6.0.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
Markdown==3.3.6
msgpack==1.0.3
Pillow==8.4.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
zope.interface==5.4.0
